Added implementation of file download from box.
This was a bit problematic. The Box API issues a redirect (302) when a
file download is requested. The redirect takes you to where the file is
actually located. The problem is that the 'request' package deal with
the redirect, but the pipe, pipe pipes the body of the redirect (empty
content), then doesn't do anything with the redirect. Had to resort to
curl which when the -L is specifed handles the redirect famously.